Update components code and test case

+++ src/fondant/components/retrieve_images_from_faiss_index/src/main.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import typing as t
 
 import dask.dataframe as dd
@@ -19,7 +20,7 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
 
     def __init__(  # PLR0913
         self,
-        dataset_path: str,
+        url_mapping_path: str,
         faiss_index_path: str,
         clip_model: str = "laion/CLIP-ViT-B-32-laion2B-s34B-b79K",
         num_images: int = 2,
@@ -30,23 +31,24 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
 
         # Download faiss index to local machine
-        logger.info(f"Downloading faiss index from {faiss_index_path}")
-
-        with fsspec.open(faiss_index_path, "rb") as f:
-            file_contents = f.read()
+        if not os.path.exists("faiss_index"):
+            logger.info(f"Downloading faiss index from {faiss_index_path}")
+            with fsspec.open(faiss_index_path, "rb") as f:
+                file_contents = f.read()
 
             with open("faiss_index", "wb") as out:
                 out.write(file_contents)
 
         self.search_index = faiss.read_index("faiss_index")
 
-        dataset = dd.read_parquet(dataset_path)
+        dataset = dd.read_parquet(url_mapping_path)
         if "url" not in dataset.columns:
             msg = "Dataset does not contain column 'url'"
             raise ValueError(msg)
         self.image_urls = dataset["url"].compute().to_list()
 
     def setup(self) -> Client:
+        """Setup LocalCudaCluster if gpu is available."""
         if self.device == "cuda":
             cluster = LocalCUDACluster()
             return Client(cluster)
@@ -54,13 +56,14 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
         return super().setup()
 
     def embed_prompt(self, prompt: str):
+        """Embed prompt using CLIP model."""
         worker = get_worker()
-        if hasattr(worker, "model"):
+        if worker and hasattr(worker, "model"):
             tokenizer = worker.tokenizer
             model = worker.model
         else:
             logger.info("Initializing model '%s' on worker '%s", self.model_id, worker)
-            tokenizer = AutoTokenizer.from_pretrained(self.model_id).to(self.device)
+            tokenizer = AutoTokenizer.from_pretrained(self.model_id)
             model = CLIPTextModelWithProjection.from_pretrained(self.model_id).to(
                 self.device,
             )
@@ -69,6 +72,7 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
             worker.tokenizer = tokenizer
 
         inputs = tokenizer([prompt], padding=True, return_tensors="pt")
+        inputs = inputs.to(self.device)
         outputs = model(**inputs)
         return outputs.text_embeds.cpu().detach().numpy().astype("float64")
 
@@ -77,10 +81,12 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
         query: float,
         number_of_images: int = 2,
     ) -> t.List[str]:
+        """Retrieve images from faiss index."""
         _, indices = self.search_index.search(query, number_of_images)
         return indices.tolist()[0]
 
     def transform_partition(self, dataframe: pd.DataFrame) -> pd.DataFrame:
+        """Transform partition of dataframe."""
         results = []
 
         for index, row in dataframe.iterrows():
@@ -112,4 +118,5 @@ class RetrieveImagesFromFaissIndex(DaskTransformComponent):
         return results_df
 
     def transform(self, dataframe: dd.DataFrame) -> dd.DataFrame:
+        """Transform dataframe."""
         return dataframe.map_partitions(self.transform_partition)
